refactor vtable module
