+++ vortex-array/src/arrays/fixed_size_list/vtable/array.rs
@@ -1,0 +1,22 @@
+// SPDX-License-Identifier: Apache-2.0
+// SPDX-FileCopyrightText: Copyright the Vortex contributors
+
+use vortex_dtype::DType;
+
+use crate::arrays::{FixedSizeListArray, FixedSizeListVTable};
+use crate::stats::StatsSetRef;
+use crate::vtable::ArrayVTable;
+
+impl ArrayVTable<FixedSizeListVTable> for FixedSizeListVTable {
+    fn len(array: &FixedSizeListArray) -> usize {
+        array.len
+    }
+
+    fn dtype(array: &FixedSizeListArray) -> &DType {
+        &array.dtype
+    }
+
+    fn stats(array: &FixedSizeListArray) -> StatsSetRef<'_> {
+        array.stats_set.to_ref(array.as_ref())
+    }
+}
