+++ vortex-array/src/arrays/fixed_size_list/vtable/validity.rs
@@ -1,0 +1,12 @@
+// SPDX-License-Identifier: Apache-2.0
+// SPDX-FileCopyrightText: Copyright the Vortex contributors
+
+use crate::arrays::FixedSizeListArray;
+use crate::validity::Validity;
+use crate::vtable::ValidityHelper;
+
+impl ValidityHelper for FixedSizeListArray {
+    fn validity(&self) -> &Validity {
+        &self.validity
+    }
+}
