+++ vortex-array/src/arrays/mod.rs
@@ -29,8 +29,8 @@ pub use constant::*;
 pub use datetime::*;
 pub use decimal::*;
 pub use extension::*;
-pub use list::*;
 pub use fixed_size_list::*;
+pub use list::*;
 pub use null::*;
 pub use primitive::*;
 pub use struct_::*;
